fix: broken imports from state/ restructure, clean up test mocks
- Fix utils/ and test imports referencing old state/types paths
  (now state/store/types after restructure)
- FeatureBar + Presets tests: use url/signedFetch props, remove
  inline mockConnection
