+++ app/components/.client/ImageViewer/components/FeatureBar/__tests__/FeatureBar.test.tsx
@@ -4,28 +4,17 @@ import { createRoutesStub } from "react-router";
 import { ViewerStoreProvider } from "../../../state/store/ViewerStoreContext";
 import { FeatureBar } from "../FeatureBar";
 
-const mockConnection = {
-  credentials: {
-    AccessKeyId: "AKIAIOSFODNN7EXAMPLE",
-    SecretAccessKey: "wJalrXUtnFEMI/K7MDENG/bPxRfiCYEXAMPLEKEY",
-    SessionToken: "token",
-    Expiration: new Date(),
-  },
-  connectionConfig: {
-    id: 1, name: "test", bucketName: "vericura-image-data", ownerScope: "org",
-    createdBy: "user", provider: "aws", endpoint: "", roleArn: null, region: "us-east-1", prefix: "",
-  },
-};
+const mockSignedFetch = vi.fn();
 
 describe("FeatureBar", () => {
-  test("renders all main sections", () => {
+  test("renders the toolbar with tablist", () => {
     const RemixStub = createRoutesStub([
       {
-        path: "/connections/vericura-image-data/lab/USL-2024-58461-31.ome.tif",
+        path: "/connections/test/lab/image.ome.tif",
         Component: () => (
           <ViewerStoreProvider
-            connection={mockConnection}
-            pathName="lab/USL-2024-58461-31.ome.tif"
+            url="https://bucket.s3.eu-central-1.amazonaws.com/lab/image.ome.tif"
+            signedFetch={mockSignedFetch}
           >
             <FeatureBar />
           </ViewerStoreProvider>
@@ -35,24 +24,11 @@ describe("FeatureBar", () => {
 
     render(
       <RemixStub
-        initialEntries={[
-          "/connections/vericura-image-data/lab/USL-2024-58461-31.ome.tif",
-        ]}
+        initialEntries={["/connections/test/lab/image.ome.tif"]}
       />
     );
 
     expect(screen.getByRole("toolbar")).toBeInTheDocument();
-
-    expect(
-      screen.getByRole("button", { name: "Channels" })
-    ).toBeInTheDocument();
-
-    expect(
-      screen.getByRole("button", { name: "Overlays" })
-    ).toBeInTheDocument();
-
-    expect(
-      screen.getByRole("button", { name: "Toggle Feature Bar" })
-    ).toBeInTheDocument();
+    expect(screen.getByRole("tablist")).toBeInTheDocument();
   });
 });
